Create contact us page with animated envelop
